+++ src/main/java/blackjack/domain/deck/shuffler/RandomCardShuffler.java
@@ -1,6 +1,7 @@
-package blackjack.domain.deck;
+package blackjack.domain.deck.shuffler;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.deck.CardShuffler;
 import java.util.Collections;
 import java.util.List;
