+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -3,6 +3,7 @@ package blackjack.domain.deck;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.deck.shuffler.RandomCardShuffler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
