+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -6,7 +6,7 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.Suit;
 import blackjack.domain.deck.Deck;
-import blackjack.domain.deck.RandomCardShuffler;
+import blackjack.domain.deck.shuffler.RandomCardShuffler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
